+++ R/leaf_sf.R
@@ -5,7 +5,7 @@
 #' @param data An sf object of geometry type point/multipoint, linestring/multilinestring or polygon/multipolygon geometry type. Required input.
 #' @param popup TRUE or FALSE of whether to have a popup.
 #' @param popup_vars_vctr Vector of quoted variable names to include in the popup. If NULL, defaults to making a leafpop::popupTable of all columns.
-#' @param popup_numeric_format A function to format all numeric variables within the tooltip text column. Defaults to adding a comma seperator. Use function(x) x to leave as is.
+#' @param popup_numeric_format A function to format all numeric variables within the popup column. Defaults to adding a comma seperator. Use function(x) x to leave as is.
 #' @param popup_vars_rename Function to rename column names for the popup. Defaults to snakecase::to_sentence_case. Use function(x) x to leave column names untransformed.
 #' @param pal Character vector of hex codes.
 #' @param alpha_point The opacity of the points. 
@@ -226,7 +226,7 @@ leaf_sf <- function(data,
 #' @param label_var Unquoted variable to label the features by. If NULL, defaults to using the colour variable.
 #' @param popup TRUE or FALSE of whether to have a popup.
 #' @param popup_vars_vctr Vector of quoted variable names to include in the popup. If NULL, defaults to making a leafpop::popupTable of all columns.
-#' @param popup_numeric_format A function to format all numeric variables within the tooltip text column. Defaults to adding a comma seperator. Use function(x) x to leave as is.
+#' @param popup_numeric_format A function to format all numeric variables within the popup column. Defaults to adding a comma seperator. Use function(x) x to leave as is.
 #' @param popup_vars_rename Function to rename column names for the popup. Defaults to snakecase::to_sentence_case. Use function(x) x to leave column names untransformed.
 #' @param pal Character vector of hex codes. 
 #' @param pal_na The hex code or name of the NA colour to be used.
@@ -245,6 +245,7 @@ leaf_sf <- function(data,
 #' @param col_method The method of colouring features, either "bin", "quantile", "continuous", or "category." If numeric, defaults to "bin".
 #' @param col_na_rm TRUE or FALSE of whether to include col_var NA values. Defaults to FALSE.
 #' @param col_title A title string that will be wrapped into the legend. 
+#' @param label_numeric_format A function to format the numeric labels. Defaults to adding a comma seperator. Use function(x) x to leave as is.
 #' @param layer_id_var Unquoted variable to be used as a shiny id, such that in the event where a feature is clicked on, the applicable value of this is available as input$map_marker_click$id or input$map_shape_click$id. 
 #' @param map_id The shiny map id for a leaflet map within a shiny app. Defaults to "map".
 #' @return A leaflet object.
@@ -300,6 +301,7 @@ leaf_sf_col <- function(data,
                         col_method = NULL,
                         col_na_rm = FALSE,
                         col_title = NULL,
+                        label_numeric_format = function(x) prettyNum(x, big.mark = ",", scientific = FALSE),
                         layer_id_var = NULL,
                         map_id = "map") {
   
@@ -336,7 +338,11 @@ leaf_sf_col <- function(data,
   
   #vectors
   col_var_vctr <- dplyr::pull(data, !!col_var)
-  label_var_vctr <- dplyr::pull(data, !!label_var)
+  
+  label_var_vctr <- data %>% 
+    dplyr::select(!!label_var) %>% 
+    dplyr::mutate_if(.predicate = is.numeric, .funs = label_numeric_format) %>% 
+    dplyr::pull(!!label_var)
 
   #logical to factor
   if (is.logical(col_var_vctr)) {